+++ .firebaserc
@@ -1,0 +1,5 @@
+{
+  "projects": {
+    "default": "contentisqueen-97ae5"
+  }
+}
