+++ .prettierrc.json
@@ -1,0 +1,5 @@
+{
+  "tabWidth": 2,
+  "semi": true,
+  "endOfLine": "auto"
+}
