+++ package.json
@@ -2,9 +2,10 @@
   "name": "creator-platform-be",
   "version": "1.0.0",
   "description": "",
-  "main": "app.js",
+  "main": "src/index.js",
   "scripts": {
-    "dev": "nodemon --exec babel-node src/app.js"
+    "dev": "nodemon --exec babel-node src/index.js",
+    "deploy": "firebase deploy --only functions:api"
   },
   "author": "Bizimungu Pascal",
   "license": "ISC",
@@ -14,6 +15,11 @@
     "@babel/polyfill": "^7.12.1",
     "@babel/preset-env": "^7.24.0",
     "@babel/register": "^7.23.7",
+    "eslint": "^8.57.0",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-n": "^16.6.2",
+    "eslint-plugin-promise": "^6.1.1",
     "nodemon": "^3.1.0"
   },
   "dependencies": {
@@ -21,14 +27,17 @@
     "@google-cloud/storage": "^7.9.0",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
-    "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
     "express": "^4.18.3",
     "express-fileupload": "^1.5.0",
     "firebase": "^10.9.0",
     "firebase-admin": "^12.0.0",
+    "firebase-functions": "^4.9.0",
     "prettier": "^3.2.5",
     "uuid": "^9.0.1"
+  },
+  "engines": {
+    "node": "18"
   }
 }
