+++ src/index.js
@@ -1,10 +1,11 @@
-import express from "express";
-import dotenv from "dotenv";
-import cors from "cors";
+const express = require("express");
+const dotenv = require("dotenv");
+const cors = require("cors");
+const functions = require("firebase-functions");
 const { initializeApp, cert } = require("firebase-admin/app");
-import fileUploader from "express-fileupload";
-import serviceAccount from "../contentisqueen-97ae5-firebase-adminsdk-qhkbo-6886ee17eb.json";
-import router from "./restful/routes";
+const fileUploader = require("express-fileupload");
+const serviceAccount = require("../contentisqueen-97ae5-firebase-adminsdk-qhkbo-6886ee17eb.json");
+const router = require("./restful/routes");
 
 dotenv.config();
 const PORT = process.env.PORT || 5000;
@@ -36,3 +37,5 @@ const start = () => {
 };
 
 start();
+
+exports.api = functions.https.onRequest(app);
