+++ src/restful/routes/opportunitiesRouters.js
@@ -1,21 +1,23 @@
-/* eslint-disable prettier/prettier */
-import { Router } from "express";
-// import protect from "../../middlewares";
-import OpportunitiesController from "../controllers/OpportunitiesControllers";
+const { Router } = require("express");
+const OpportunitiesController = require("../controllers/opportunitiesControllers");
 
 const router = Router();
 
 router.get("/", OpportunitiesController.getAllOpportunities);
 // router.get("/open", OpportunitiesController.getOpenOpportunities);
-router.get('/id/:user_id', OpportunitiesController.getAllOpportunitiesByUserId);
-router.get('/status/:status', OpportunitiesController.getOpportunitiesByStatus);
+router.get("/id/:user_id", OpportunitiesController.getAllOpportunitiesByUserId);
+router.get("/status/:status", OpportunitiesController.getOpportunitiesByStatus);
 // router.get("/in_progress", OpportunitiesController.getInProgressOpportunities);
 // router.get("/completed", OpportunitiesController.getCompletedOpportunities);
 // router.get("/:doc_type/:opportunity_id", OpportunitiesController.getOpportunityById); // Make sure this route is correct
-router.delete("/opportunityid/:opportunity_id", OpportunitiesController.deleteOpportunityById);
-router.get("/opportunityid/:opportunity_id", OpportunitiesController.getOpportunityById);
-
-
+router.delete(
+  "/opportunityid/:opportunity_id",
+  OpportunitiesController.deleteOpportunityById,
+);
+router.get(
+  "/opportunityid/:opportunity_id",
+  OpportunitiesController.getOpportunityById,
+);
 
 // GET endpoint to retrieve individual opportunity by ID
 // router.get("/:doc_type/:opportunity_id", OpportunitiesController.getOpportunityById);
@@ -25,33 +27,40 @@ router.get("/opportunityid/:opportunity_id", OpportunitiesController.getOpportun
 
 // POST endpoint for creating opportunities
 router.post("/", async (req, res) => {
-    const { type } = req.body;
-    if (!type || !['job', 'pitch', 'campaign'].includes(type)) {
-      return res.status(400).json({ message: "Invalid or missing opportunity type" });
+  const { type } = req.body;
+  if (!type || !["job", "pitch", "campaign"].includes(type)) {
+    return res
+      .status(400)
+      .json({ message: "Invalid or missing opportunity type" });
+  }
+
+  try {
+    switch (type) {
+      case "job":
+        return OpportunitiesController.createJobOpportunity(req, res);
+      case "pitch":
+        return OpportunitiesController.createPitchOpportunity(req, res);
+      case "campaign":
+        return OpportunitiesController.createCampaignOpportunity(req, res);
+      default:
+        return res.status(400).json({ message: "Invalid opportunity type" });
     }
-  
-    try {
-      switch (type) {
-        case 'job':
-          return OpportunitiesController.createJobOpportunity(req, res);
-        case 'pitch':
-          return OpportunitiesController.createPitchOpportunity(req, res);
-        case 'campaign':
-          return OpportunitiesController.createCampaignOpportunity(req, res);
-        default:
-          return res.status(400).json({ message: "Invalid opportunity type" });
-      }
-    } catch (error) {
-      console.error(error);
-      return res.status(500).json({ message: "Server error" });
-    }
-  });
+  } catch (error) {
+    console.error(error);
+    return res.status(500).json({ message: "Server error" });
+  }
+});
 
-  
-//DELETE endpoint
-router.delete("/:doc_type/:opportunity_id", OpportunitiesController.deleteOpportunityById); // New route for delete
+// DELETE endpoint
+router.delete(
+  "/:doc_type/:opportunity_id",
+  OpportunitiesController.deleteOpportunityById,
+); // New route for delete
 
-//UPDATE endpoint
-router.put("/:doc_type/:opportunity_id", OpportunitiesController.updateOpportunityById); // New route for update
+// UPDATE endpoint
+router.put(
+  "/:doc_type/:opportunity_id",
+  OpportunitiesController.updateOpportunityById,
+); // New route for update
 
-module.exports = { router };
+exports.router = router;
